% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazygrow.R
\name{lazygrow}
\alias{lazygrow}
\title{Quick easigrow test}
\usage{
lazygrow(scale, sequence, res_filename)
}
\arguments{
\item{scale}{Peak stress in MPa}

\item{sequence}{Sequence to apply}

\item{res_filename}{Name to give the text file containing easigrow results}
}
\value{
A tibble with the block and crack growth sizes
}
\description{
lazygrow
}
