% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_easigrow.R
\name{read_easigrow}
\alias{read_easigrow}
\title{Read easigrow file}
\usage{
read_easigrow(filename)
}
\arguments{
\item{filename}{The name of the output file from easigrow}
}
\value{
A tibble with the block and crack growth sizes
}
\description{
Reads the output from easigrow saved to the specified file
}
\concept{easigrow}
