% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_easigrow.R
\name{run_easigrow}
\alias{run_easigrow}
\title{Easigrow}
\usage{
run_easigrow(filename, scale, flags = NULL, options = NULL)
}
\arguments{
\item{filename}{Text file with normalised stress spectrum}

\item{scale}{Peak stress value for the spectrum (MPa)}

\item{flags}{String containing flag specifiers for the simulation}

\item{options}{String containing option specifiers for the simulation}
}
\description{
Executes easigrow for crack growth measurement
}
