% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_easigrow.R
\name{write_easigrow}
\alias{write_easigrow}
\alias{write_easigrow.data.frame}
\title{Write easigrow text file}
\usage{
write_easigrow(x, filename, ...)

\method{write_easigrow}{data.frame}(x, filename, ...)
}
\arguments{
\item{x}{Input data}

\item{filename}{Name of text file to write to}

\item{...}{Numeric or character arguments}
}
\value{
NULL, writes data to the specified filename
}
\description{
write_easigrow
}
